Some changes start of multi windows
