+++ src/App/Settings.java
@@ -17,6 +17,7 @@ public class Settings {
     private final int subImageHeight = 66;
     private final int maxCardsOnTable = 7;
     private final int autoPlayStepsPerSecond = 10;
+    private final String themeName = "cards";
 
     public int getWidth() {
         return this.width;
@@ -71,5 +72,9 @@ public class Settings {
     }
     public int getAutoPlayStepsPerSecond() {
         return this.autoPlayStepsPerSecond;
+    }
+
+    public String getThemeName() {
+        return themeName;
     }
 }
